compiledb: Remove the VVI tag
Instead of just picking one of either VVB or VVI, it now just remaps VVI
to VVB. The EntryFilter code has been replaced with a more general
filter that avoids entries that have all the same details except the
frequency. When there is a duplicate the highest frequency is used. This
also fixes a problem that the variants for verbs were accidentally
getting removed, like for “chance”.